Fixed issue: Mutiple themes in production environment
When multiple themes are enabled and Gleez environment is set to production, the file paths gets cached, so the templates or views also cached and returns the cached view instead of the active theme view, so theme based caching is necessary to cache for performance. Still it has rough edges, need a better way in future.