New feature: Added Mango class (new version). WIP
It's a complete class replacement for Mango Reader module.

   Warning: TEMPORARY DO NOT USE Mango Reader module!!!

This class will be used to for full access to the sessions, logs,
and other data to improve performance. I have provided a detailed
description of the class in the form of PHPDoc comments.
Please read them initially.

Todo: Mango_Collection
      Mango_Document
      GridFS support

      Log Reader/Writer
      Session Reader/Writer
      Cache Reader/Writer